Refactor: Split App.cpp into smaller modules and add AppUtil - Split App.cpp into AppStart.cpp, AppUpdate.cpp, and AppEnd.cpp for improved code organization and maintainability - Added AppUtil.hpp and AppUtil.cpp to encapsulate utility functions related to the App class
