+++ include/AppUtil.hpp
@@ -1,0 +1,4 @@
+#ifndef APP_UTIL_HPP
+#define APP_UTIL_HPP
+
+#endif //APP_UTIL_HPP
