--- src/App.cpp
@@ -1,21 +1,0 @@
-#include "App.hpp"
-
-#include "Util/Image.hpp"
-#include "Util/Input.hpp"
-#include "Util/Keycode.hpp"
-#include "Util/Logger.hpp"
-
-void App::Start() {
-    LOG_TRACE("Start");
-
-    m_CurrentState = State::UPDATE;
-}
-
-void App::Update() {
-
-    m_Root.Update();
-}
-
-void App::End() { // NOLINT(this method will mutate members in the future)
-    LOG_TRACE("End");
-}
