+++ src/AppEnd.cpp
@@ -1,0 +1,6 @@
+#include "App.hpp"
+#include"Util/Logger.hpp"
+
+void App::End() { // NOLINT(this method will mutate members in the future)
+    LOG_TRACE("End");
+}
