+++ src/AppUpdate.cpp
@@ -1,0 +1,16 @@
+#include "App.hpp"
+
+#include "Util/Input.hpp"
+#include "Util/Keycode.hpp"
+#include "Util/Logger.hpp"
+
+
+
+void App::Update() {
+
+    if (Util::Input::IsKeyPressed(Util::Keycode::ESCAPE) ||
+        Util::Input::IfExit()) {
+        m_CurrentState = State::END;
+    }
+    m_Root.Update();
+}
